target/riscv: Only set icount.PENDING bit when icount trigger matches
Refer to RISC-V debug specification:
- When count is 1 and the trigger matches, then pending becomes set.
- When pending is set, the trigger fires just before any further
  instructions are executed in a mode where the trigger is enabled.

This commit adds helper function to set the pending bit of icount
trigger and calls the function when the icount trigger count is 1 and it
matches. Also we remove icount trigger fire in helper_itrigger_match(),
since this function is called when any instruction retires. This is not
correct timing to fire the pending icount trigger.

Signed-off-by: Alvin Chang <alvinga@andestech.com>