target/riscv: Fix index of fired breakpoint and watchpoint
In current implementation, the index of fired breakpoint and watchpoint
is hard-code as DBG_ACTION_BP, which is 0. This is a bug because we may
match/fire breakpoints and watchpoints at other indexes, i.e., index 0~7
since we have 8 triggers. Fix it by implementing two functions to
iterate the cpu_breakpoint and the cpu_watchpoint to compare specific
information. The index of the cpu_breakpoint and the cpu_watchpoint is
the index we want to execute corresponding trigger action.

Signed-off-by: Alvin Chang <alvinga@andestech.com>